requirements.txt
anyio            ==  4.5.2  
annotated-types  ==  0.7.0  
bidict           ==  0.23.1 
click            ==  8.1.7  
exceptiongroup   ==  1.2.2  
fastapi          ==  0.115.5
h11              ==  0.14.0 
idna             ==  3.10   
pip              ==  20.0.2 
pkg-resources    ==  0.0.0  
pydantic         ==  2.9.2  
pydantic-core    ==  2.23.4 
python-engineio  ==  4.10.1 
python-socketio  ==  5.11.4 
setuptools       ==  44.0.0 
simple-websocket ==  1.1.0  
sniffio          ==  1.3.1  
starlette        ==  0.41.2 
typing-extensions==  4.12.2 
uvicorn          ==  0.32.0 
wsproto          ==  1.2.0  